Moved wedding cake to examples and adjusted tests
