+++ anesthetic/examples/_matplotlib_agg.py
@@ -1,0 +1,3 @@
+"""When imported forces matplotlib to use Agg backend. Useful for tests."""
+import matplotlib
+matplotlib.use('Agg')
