--- tests/matplotlib_agg.py
@@ -1,2 +1,0 @@
-import matplotlib
-matplotlib.use('Agg')
