+++ tests/test_examples.py
@@ -4,9 +4,10 @@ from anesthetic.examples.utils import (
     random_ellipsoid, random_covariance, volume_n_ball, log_volume_n_ball
 )
 from anesthetic.examples.perfect_ns import (
-    gaussian, correlated_gaussian
+    gaussian, correlated_gaussian, wedding_cake
 )
 from scipy.spatial import ConvexHull
+from scipy.special import logsumexp
 
 
 def test_random_covariance():
@@ -81,3 +82,24 @@ def test_perfect_ns_correlated_gaussian():
     assert (samples[:-nlive].nlive >= nlive).all()
     assert_allclose(samples[np.arange(ndims)].mean(), mean, atol=1e-2)
     assert_allclose(samples[np.arange(ndims)].cov(), cov/(ndims+2), atol=1e-2)
+
+
+def test_wedding_cake():
+    np.random.seed(0)
+    nlive = 500
+    ndims = 4
+    sigma = 0.01
+    alpha = 0.5
+    samples = wedding_cake(nlive, ndims, sigma, alpha)
+
+    assert samples.nlive.iloc[0] == nlive
+    assert samples.nlive.iloc[-1] == 1
+    assert (samples.nlive <= nlive).all()
+
+    mean = np.sqrt(ndims/2) * (np.log(4*ndims*sigma**2)-1) / np.log(alpha)
+    std = -np.sqrt(ndims/2)/np.log(alpha)
+    i = np.arange(mean + std*10)
+    logZ = logsumexp(-alpha**(2*i/ndims)/8/sigma**2
+                     + i*np.log(alpha) + np.log(1-alpha))
+
+    assert_allclose(samples.logZ(), logZ, atol=3*samples.logZ(12).std())
