+++ tests/test_plot.py
@@ -1,4 +1,4 @@
-import matplotlib_agg  # noqa: F401
+import anesthetic.examples._matplotlib_agg  # noqa: F401
 import pytest
 import numpy as np
 import sys
